refactor: split cross surface builders
