+++ benchbox/core/equivalence/cross_surface.py
@@ -73,6 +73,16 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
+from benchbox.core.equivalence.builders import (
+    CrossSurfaceData,
+    build_amplab_duckdb,
+    build_clickbench_duckdb,
+    build_coffeeshop_duckdb,
+    build_h2odb_duckdb,
+    build_joinorder_synthetic_duckdb,
+    build_read_primitives_duckdb,
+    build_ssb_duckdb,
+)
 from benchbox.core.equivalence.dataframe_surface import (
     DATAFRAME_BACKENDS,
     SurfaceDivergence,
@@ -314,27 +324,6 @@ def _resolve_order_term(
         if index is not None:
             return index
     return expr_to_index.get(target.sql(dialect="duckdb", normalize=True).lower())
-
-
-@dataclass(frozen=True)
-class CrossSurfaceData:
-    """Everything the gate needs to compare a benchmark's two surfaces.
-
-    Returned by a benchmark's ``build`` function on a freshly generated, loaded
-    DuckDB cell. ``reference_sql`` and ``dataframe_query`` are keyed by the SAME
-    query id (the two surfaces are confirmed to correspond 1:1).
-    """
-
-    connection: Any
-    query_ids: Sequence[Any]
-    reference_sql: Callable[[Any], str]
-    dataframe_query: Callable[[Any], Any]
-    # The benchmark instance + the directory its data was generated into. The
-    # DataFrame surface is loaded from these via the real production loader (see
-    # :func:`build_production_contexts`), reading the SAME generated files the
-    # DuckDB SQL reference loaded, so the comparison stays a single bounded cell.
-    benchmark: Any
-    data_dir: Path
 
 
 @dataclass(frozen=True)
@@ -601,321 +590,6 @@ def build_production_contexts(
         adapter = adapter_cls()
         contexts[backend] = adapter.load_benchmark_into_context(benchmark, Path(data_dir), scale_factor=scale_factor)
     return contexts
-
-
-def _load_duckdb_cell(benchmark: Any, output_dir: Path, table_names: Sequence[str], *, label: str) -> Any:
-    """Create an in-memory DuckDB, build the schema, load the data, and verify it.
-
-    Shared by every benchmark's ``build`` function: it runs the benchmark's
-    create-table DDL, loads the generated files via the real
-    :class:`~benchbox.platforms.duckdb.DuckDBAdapter`, and asserts each expected
-    table got rows. A silent empty/partial load would make every query compare
-    empty-vs-empty and report a FALSE green, so an empty table fails loudly here.
-    Returns the populated connection (closed by the caller on any later error).
-    ``duckdb``/adapter imports are deferred so importing this module stays cheap.
-    """
-    import duckdb
-
-    from benchbox.platforms.duckdb import DuckDBAdapter
-
-    connection = duckdb.connect(":memory:")
-    try:
-        for statement in benchmark.get_create_tables_sql(dialect="duckdb").strip().split(";"):
-            if statement.strip():
-                connection.execute(statement.strip())
-        table_stats, _, _ = DuckDBAdapter(database=":memory:").load_data(benchmark, connection, Path(output_dir))
-
-        stats_lower = {str(name).lower(): rows for name, rows in table_stats.items()}
-        empty = [name for name in table_names if stats_lower.get(name.lower(), 0) <= 0]
-        if empty:
-            raise RuntimeError(f"{label} load failed - no rows in {empty} (stats={table_stats})")
-    except Exception:
-        connection.close()
-        raise
-    return connection
-
-
-def build_ssb_duckdb(scale_factor: float, output_dir: Path) -> CrossSurfaceData:
-    """Generate SSB data, load it into in-memory DuckDB, and wire both surfaces.
-
-    Platform/generator imports are deferred so importing this module stays cheap.
-    """
-    from benchbox.core.ssb.benchmark import SSBBenchmark
-    from benchbox.core.ssb.dataframe_queries import SSB_DATAFRAME_QUERIES
-    from benchbox.core.ssb.generator import SSBDataGenerator
-    from benchbox.core.ssb.schema import TABLES
-
-    output_dir = Path(output_dir)
-    SSBDataGenerator(scale_factor=scale_factor, output_dir=output_dir).generate_data()
-    benchmark = SSBBenchmark(scale_factor=scale_factor, output_dir=output_dir)
-
-    connection = _load_duckdb_cell(benchmark, output_dir, [table["name"] for table in TABLES.values()], label="SSB")
-    return CrossSurfaceData(
-        connection=connection,
-        query_ids=list(benchmark.get_queries().keys()),
-        reference_sql=lambda query_id: benchmark.get_query(query_id),
-        dataframe_query=lambda query_id: SSB_DATAFRAME_QUERIES.get_or_raise(query_id),
-        benchmark=benchmark,
-        data_dir=output_dir,
-    )
-
-
-def build_amplab_duckdb(scale_factor: float, output_dir: Path) -> CrossSurfaceData:
-    """Generate AMPLab data, load it into in-memory DuckDB, and wire both surfaces.
-
-    The SQL surface keys queries by ``"1"``, ``"1a"``, ``"2"`` .. ``"5"`` (8 ids)
-    while the DataFrame registry keys the SAME logical queries by the ``Q``-prefixed
-    ``"Q1"``, ``"Q1a"``, ``"Q2"`` .. ``"Q5"`` (1:1, mechanical prefix), so the
-    DataFrame accessor maps the SQL id across. Platform/generator imports are
-    deferred so importing this module stays cheap.
-    """
-    from benchbox.core.amplab.benchmark import AMPLabBenchmark
-    from benchbox.core.amplab.dataframe_queries import AMPLAB_DATAFRAME_QUERIES
-    from benchbox.core.amplab.generator import AMPLabDataGenerator
-    from benchbox.core.amplab.schema import TABLES
-
-    output_dir = Path(output_dir)
-    AMPLabDataGenerator(scale_factor=scale_factor, output_dir=output_dir).generate_data()
-    benchmark = AMPLabBenchmark(scale_factor=scale_factor, output_dir=output_dir)
-
-    connection = _load_duckdb_cell(benchmark, output_dir, [table["name"] for table in TABLES.values()], label="AMPLab")
-    return CrossSurfaceData(
-        connection=connection,
-        query_ids=list(benchmark.get_queries().keys()),
-        reference_sql=lambda query_id: benchmark.get_query(query_id),
-        dataframe_query=lambda query_id: AMPLAB_DATAFRAME_QUERIES.get_or_raise(f"Q{query_id}"),
-        benchmark=benchmark,
-        data_dir=output_dir,
-    )
-
-
-def build_clickbench_duckdb(scale_factor: float, output_dir: Path) -> CrossSurfaceData:
-    """Generate ClickBench data, load it into in-memory DuckDB, and wire both surfaces.
-
-    Mirrors :func:`build_ssb_duckdb`. ClickBench is a single wide table (``hits``);
-    the SQL surface (ids ``Q1``..``Q43``) and the DataFrame registry key by the same
-    ids 1:1. Imports are deferred so importing this module stays cheap.
-    """
-    import duckdb
-
-    from benchbox.core.clickbench.benchmark import ClickBenchBenchmark
-    from benchbox.core.clickbench.dataframe_queries import CLICKBENCH_DATAFRAME_QUERIES
-    from benchbox.core.clickbench.generator import ClickBenchDataGenerator
-    from benchbox.core.clickbench.schema import TABLES
-    from benchbox.platforms.duckdb import DuckDBAdapter
-
-    output_dir = Path(output_dir)
-    ClickBenchDataGenerator(scale_factor=scale_factor, output_dir=output_dir).generate_data()
-    benchmark = ClickBenchBenchmark(scale_factor=scale_factor, output_dir=output_dir)
-
-    connection = duckdb.connect(":memory:")
-    try:
-        for statement in benchmark.get_create_tables_sql(dialect="duckdb").strip().split(";"):
-            if statement.strip():
-                connection.execute(statement.strip())
-        table_stats, _, _ = DuckDBAdapter(database=":memory:").load_data(benchmark, connection, output_dir)
-
-        # A silent empty/partial load would compare empty-vs-empty and report a
-        # FALSE green, so verify the table loaded.
-        table_names = [table["name"] for table in TABLES.values()]
-        stats_lower = {str(name).lower(): rows for name, rows in table_stats.items()}
-        empty = [name for name in table_names if stats_lower.get(name.lower(), 0) <= 0]
-        if empty:
-            raise RuntimeError(f"ClickBench load failed - no rows in {empty} (stats={table_stats})")
-    except Exception:
-        connection.close()
-        raise
-
-    return CrossSurfaceData(
-        connection=connection,
-        query_ids=list(benchmark.get_queries().keys()),
-        reference_sql=lambda query_id: benchmark.get_query(query_id),
-        dataframe_query=lambda query_id: CLICKBENCH_DATAFRAME_QUERIES.get_or_raise(query_id),
-        benchmark=benchmark,
-        data_dir=output_dir,
-    )
-
-
-def build_coffeeshop_duckdb(scale_factor: float, output_dir: Path) -> CrossSurfaceData:
-    """Generate CoffeeShop data, load it into in-memory DuckDB, and wire both surfaces.
-
-    Platform/generator imports are deferred so importing this module stays cheap.
-    """
-    import duckdb
-
-    from benchbox.core.coffeeshop.benchmark import CoffeeShopBenchmark
-    from benchbox.core.coffeeshop.dataframe_queries import COFFEESHOP_DATAFRAME_QUERIES
-    from benchbox.core.coffeeshop.generator import CoffeeShopDataGenerator
-    from benchbox.core.coffeeshop.schema import TABLES
-    from benchbox.platforms.duckdb import DuckDBAdapter
-
-    output_dir = Path(output_dir)
-    CoffeeShopDataGenerator(scale_factor=scale_factor, output_dir=output_dir).generate_data()
-    benchmark = CoffeeShopBenchmark(scale_factor=scale_factor, output_dir=output_dir)
-
-    connection = duckdb.connect(":memory:")
-    try:
-        for statement in benchmark.get_create_tables_sql(dialect="duckdb").strip().split(";"):
-            if statement.strip():
-                connection.execute(statement.strip())
-        table_stats, _, _ = DuckDBAdapter(database=":memory:").load_data(benchmark, connection, output_dir)
-
-        # A silent empty/partial load would make every query compare
-        # empty-vs-empty and report a FALSE green, so verify each table loaded.
-        table_names = [table["name"] for table in TABLES.values()]
-        stats_lower = {str(name).lower(): rows for name, rows in table_stats.items()}
-        empty = [name for name in table_names if stats_lower.get(name.lower(), 0) <= 0]
-        if empty:
-            raise RuntimeError(f"CoffeeShop load failed - no rows in {empty} (stats={table_stats})")
-    except Exception:
-        connection.close()
-        raise
-
-    return CrossSurfaceData(
-        connection=connection,
-        query_ids=list(benchmark.get_queries().keys()),
-        reference_sql=lambda query_id: benchmark.get_query(query_id),
-        dataframe_query=lambda query_id: COFFEESHOP_DATAFRAME_QUERIES.get_or_raise(query_id),
-        benchmark=benchmark,
-        data_dir=output_dir,
-    )
-
-
-def build_joinorder_synthetic_duckdb(scale_factor: float, output_dir: Path) -> CrossSurfaceData:
-    """Generate synthetic Join Order data, load into DuckDB, and wire both surfaces.
-
-    Mirrors :func:`build_clickbench_duckdb`. The SQL surface and the
-    ``JOINORDER_DATAFRAME_QUERIES`` registry key by the same 13 ids (e.g. ``1a``,
-    ``2a``). The generator is seeded, so the bounded cell is reproducible.
-    """
-    import duckdb
-
-    from benchbox.core.joinorder_synthetic.benchmark import JoinOrderSyntheticBenchmark
-    from benchbox.core.joinorder_synthetic.dataframe_queries import JOINORDER_DATAFRAME_QUERIES
-    from benchbox.core.joinorder_synthetic.generator import JoinOrderGenerator
-    from benchbox.platforms.duckdb import DuckDBAdapter
-
-    output_dir = Path(output_dir)
-    JoinOrderGenerator(scale_factor=scale_factor, output_dir=output_dir).generate_data()
-    benchmark = JoinOrderSyntheticBenchmark(scale_factor=scale_factor, output_dir=output_dir)
-
-    connection = duckdb.connect(":memory:")
-    try:
-        for statement in benchmark.get_create_tables_sql(dialect="duckdb").strip().split(";"):
-            if statement.strip():
-                connection.execute(statement.strip())
-        table_stats, _, _ = DuckDBAdapter(database=":memory:").load_data(benchmark, connection, output_dir)
-
-        # A silent empty/partial load would compare empty-vs-empty and report a
-        # FALSE green, so verify each table loaded.
-        stats_lower = {str(name).lower(): rows for name, rows in table_stats.items()}
-        empty = [name for name in benchmark.get_table_names() if stats_lower.get(name.lower(), 0) <= 0]
-        if empty:
-            raise RuntimeError(f"JoinOrder synthetic load failed - no rows in {empty} (stats={table_stats})")
-    except Exception:
-        connection.close()
-        raise
-
-    return CrossSurfaceData(
-        connection=connection,
-        query_ids=list(benchmark.get_queries().keys()),
-        reference_sql=lambda query_id: benchmark.get_query(query_id),
-        dataframe_query=lambda query_id: JOINORDER_DATAFRAME_QUERIES.get_or_raise(query_id),
-        benchmark=benchmark,
-        data_dir=output_dir,
-    )
-
-
-def build_h2odb_duckdb(scale_factor: float, output_dir: Path) -> CrossSurfaceData:
-    """Generate H2O-DB data, load it into in-memory DuckDB, and wire both surfaces.
-
-    Mirrors :func:`build_ssb_duckdb` via :func:`_load_duckdb_cell`. H2O-DB is a
-    single wide ``trips`` table; the SQL surface (``H2OQueryManager``, ids
-    ``Q1``..``Q10``) and the ``H2ODB_DATAFRAME_QUERIES`` registry key by the same
-    ids 1:1, so no id normalization is needed. The generator is seeded
-    (``random.seed(42)`` in its ``__init__``), so the bounded cell is reproducible.
-
-    H2O-DB's generator base is the published 10M-row small tier, so the shared
-    ``EQUIVALENCE_SCALE`` (0.1) would generate ~1M rows - far from a cheap bounded
-    cell and ~100x the other gates' row counts. The gate therefore runs at a
-    smaller per-gate ``scale_factor`` (see :data:`GATES`); every query stays fully
-    discriminating at that cell (low-cardinality group keys: passenger_count <= 6,
-    vendor_id = 2, hour = 24, year span 5, pickup_location_id <= 263; the
-    percentile and top-10 queries are well-determined over the remaining rows).
-    Platform/generator imports are deferred so importing this module stays cheap.
-    """
-    from benchbox.core.h2odb.benchmark import H2OBenchmark
-    from benchbox.core.h2odb.dataframe_queries import H2ODB_DATAFRAME_QUERIES
-    from benchbox.core.h2odb.generator import H2ODataGenerator
-    from benchbox.core.h2odb.schema import TABLES
-
-    output_dir = Path(output_dir)
-    H2ODataGenerator(scale_factor=scale_factor, output_dir=output_dir).generate_data()
-    benchmark = H2OBenchmark(scale_factor=scale_factor, output_dir=output_dir)
-
-    connection = _load_duckdb_cell(benchmark, output_dir, [table["name"] for table in TABLES.values()], label="H2O-DB")
-    return CrossSurfaceData(
-        connection=connection,
-        query_ids=list(benchmark.get_queries().keys()),
-        reference_sql=lambda query_id: benchmark.get_query(query_id),
-        dataframe_query=lambda query_id: H2ODB_DATAFRAME_QUERIES.get_or_raise(query_id),
-        benchmark=benchmark,
-        data_dir=output_dir,
-    )
-
-
-def build_read_primitives_duckdb(scale_factor: float, output_dir: Path) -> CrossSurfaceData:
-    """Generate Read Primitives data, load it into DuckDB, and wire both surfaces.
-
-    Read Primitives reuses the canonical TPC-H schema (8 tables) and dbgen-based
-    generator, so the bounded cell is reproducible for a given scale. Unlike
-    ssb/h2odb - whose canonical SQL is already DuckDB-native - Read Primitives
-    ships a rich per-dialect *catalog-variant* system: its canonical SQL uses
-    portable/ANSI forms (``TRANSFORM``, ``STRUCT(...)``, ``APPROX_*``) that DuckDB
-    does not accept verbatim. The trusted reference here is therefore the
-    DuckDB-DIALECT surface - ``benchmark.get_queries(dialect="duckdb")``, which
-    returns each query's authored DuckDB catalog variant verbatim and translates
-    the rest - NOT the raw canonical SQL, so the comparison runs the SQL a DuckDB
-    production run would actually execute.
-
-    The DataFrame registry keys 152 queries by ids that are a strict subset of the
-    157 SQL ids (the 5 SQL-only ids are the documented ``get_skip_for_dataframe``
-    set). Four of those 152 - ``fulltext_*`` (FTS, no portable DuckDB form) and
-    ``json_extract_simple`` (skipped on duckdb) - have NO DuckDB SQL surface
-    (``get_queries(dialect="duckdb")`` drops them), so there is no DuckDB reference
-    to compare and they are excluded here, leaving 148 gateable queries. The
-    exclusion is the same discipline as a missing-backend skip: there is genuinely
-    nothing to compare on this engine, not a muted divergence.
-
-    Platform/generator imports are deferred so importing this module stays cheap.
-    """
-    from benchbox.core.read_primitives.benchmark import ReadPrimitivesBenchmark
-    from benchbox.core.read_primitives.dataframe_queries import get_dataframe_queries
-    from benchbox.core.read_primitives.schema import TABLES
-
-    output_dir = Path(output_dir)
-    benchmark = ReadPrimitivesBenchmark(scale_factor=scale_factor, output_dir=output_dir)
-    benchmark.data_generator.output_dir = output_dir
-    benchmark.data_generator.generate_data()
-
-    # The DuckDB-dialect surface is the trusted reference (catalog variants
-    # verbatim + translation). Computed once; queries DuckDB cannot transpile are
-    # absent from this dict and are filtered out of the gateable id set below.
-    duckdb_sql = benchmark.get_queries(dialect="duckdb")
-    registry = get_dataframe_queries()
-    gateable_ids = [qid for qid in registry.get_query_ids() if qid in duckdb_sql]
-
-    connection = _load_duckdb_cell(
-        benchmark, output_dir, [table["name"] for table in TABLES.values()], label="Read Primitives"
-    )
-    return CrossSurfaceData(
-        connection=connection,
-        query_ids=gateable_ids,
-        reference_sql=lambda query_id: duckdb_sql[query_id],
-        dataframe_query=lambda query_id: registry.get_or_raise(query_id),
-        benchmark=benchmark,
-        data_dir=output_dir,
-    )
 
 
 # Q18 is `SELECT UserID, SearchPhrase, COUNT(*) ... GROUP BY ... LIMIT 10` with NO
